Removed a couple of warnings and changed the mps optimization test from 20 qubits to 12

+++ Simulators/MPSDummySimulator.h
@@ -578,9 +578,9 @@ class MPSDummySimulator {
 
       size_t gateCount = 0;
       for (const auto& lp : orderedLP) {
-        if (placed.size() == nrQubits) break;
+        if (placed.size() == static_cast<size_t>(nrQubits)) break;
         for (const auto& p : lp) {
-          if (placed.size() == nrQubits) break;
+          if (placed.size() == static_cast<size_t>(nrQubits)) break;
           const bool q1In = placed.count(p.q1) > 0;
           const bool q2In = placed.count(p.q2) > 0;
